requirements.txt
boto3
botocore
Flask
marshmallow
pyjwt
werkzeug
jsonschema
flask-cors
python-dotenv 